Feat: was updated a readme.md file and was added some validations

+++ src/usecases/post.usecase.js
@@ -1,5 +1,4 @@
 const Post = require('../models/post.model')
-const usersModel = require('../models/users.model')
 
 const createError = require('http-errors')
 
@@ -13,14 +12,29 @@ async function getAll() {
     return allPosts
 }
 
-async function deleteById(id) {
-    const postDeleted = await Post.findByIdAndDelete(id)
+async function deleteById(id,userId) {
+
+    const postActual = await Post.findById(id)
+    
+    if(!postActual) throw createError(404, "Post not found")
+
+    if (userId != postActual.user) {
+        throw createError(401, "You cant delete this post, you not the owner")
+    }
+
+    const postDeleted = await Post.findByIdAndDelete(id)  
+
     if (!postDeleted) throw createError(404, "Post not found")
+
     return postDeleted
+
 }
 
 async function updateByID(id, newPostData) {
     newPostData.updatedAt = Date.now()
+    const actualPost = await Post.findById(id)
+    if(actualPost.user != newPostData.user) throw createError(409, "you cant update a user")
+
     const updatedPost= await Post.findByIdAndUpdate(id, newPostData, { new: true })
     if (!updatedPost) throw createError(404, "Post not found")
